+++ src/structs.jl
@@ -103,4 +103,7 @@ function from_query(query::TidierDB.SQLQuery)
     return new_query
 end
 
+"""
+$docstring_t
+"""
 t(table) = from_query(table)